simplifying and fixed an issue in layCards
in some places, there would quite long repeated sections for each player. i simplified this by putting them in a list and using an index to refer to the right one.

also in layCards, it was checking if the player had any cards AFTER asking them to play their cards and this could lead to problems when a player played their last card and then didn't have any cards left. it seems to work better with these two switched.